template: view: add ajax based browser_lazy
This introduce new browser view with partial content.

Browsing view of booklist is not sent at once on page loading
but only a partial subset is sent. Next, when user is scrolling
screen and arrives at bottom view, an AJAX http request is sent
by the browser to get next booklist subset.

Signed-off-by: Yannick Gicquel <ygicquel@gmail.com>