bookdir: directory not scanned when containing a special file
This commit add the ability for the user to 'mark' a directory so
that the book directory scanner will pass over it and will
ignore pdf files it eventually contains.

This feature is based on the presence detection of a file named
".ebook-ignore-dir". When this empty hidden file is detected, the
directory is not scanned.

Signed-off-by: Yannick Gicquel <ygicquel@gmail.com>